# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API AlexTa",
        "description": "Documentation de l'API de l'application AlexTa",
        "contact": {
            "email": "support@alexta.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/supplements-localisation": {
            "get": {
                "tags": [
                    "SupplementLocalisation"
                ],
                "summary": "Afficher tous les suppléments de localisation",
                "operationId": "3cfabb447edc04a44935f2c276e3f55c",
                "responses": {
                    "200": {
                        "description": "Liste des suppléments de localisation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/SupplementLocalisation"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erreur interne du serveur"
                    }
                }
            },
            "post": {
                "tags": [
                    "SupplementLocalisation"
                ],
                "summary": "Créer un supplément de localisation",
                "operationId": "95201242da1c3ab2b624bb092d2fd3df",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lieu",
                                    "montant"
                                ],
                                "properties": {
                                    "lieu": {
                                        "type": "string",
                                        "example": "Paris"
                                    },
                                    "montant": {
                                        "type": "integer",
                                        "example": 15000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Supplément de localisation créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementLocalisation"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/supplements-localisation/{id}": {
            "get": {
                "tags": [
                    "SupplementLocalisation"
                ],
                "summary": "Afficher un supplément de localisation spécifique",
                "operationId": "5872601ab7aa2b8d77ba011bc1246810",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément de localisation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplément de localisation trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementLocalisation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplément de localisation non trouvé"
                    }
                }
            },
            "put": {
                "tags": [
                    "SupplementLocalisation"
                ],
                "summary": "Mettre à jour un supplément de localisation",
                "operationId": "2a893f8a7c28fa14a05992f2e93705ec",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément de localisation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "lieu": {
                                        "type": "string",
                                        "example": "Lyon"
                                    },
                                    "montant": {
                                        "type": "integer",
                                        "example": 20000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supplément de localisation mis à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SupplementLocalisation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Supplément de localisation non trouvé"
                    }
                }
            },
            "delete": {
                "tags": [
                    "SupplementLocalisation"
                ],
                "summary": "Supprimer un supplément de localisation",
                "operationId": "6980c203b4fad125c36fd63c6fa4332a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du supplément de localisation",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Supplément de localisation supprimé avec succès"
                    },
                    "404": {
                        "description": "Supplément de localisation non trouvé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SupplementLocalisation": {
                "title": "Supplement Localisation",
                "required": [
                    "lieu",
                    "montant"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "lieu": {
                        "type": "string",
                        "example": "Libreville"
                    },
                    "montant": {
                        "type": "integer",
                        "example": 5000
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}